+++ scopes/code/ui/code-compare/code-compare.tsx
@@ -45,7 +45,7 @@ export function CodeCompare({ fileIconSlot, className, CodeView = CodeCompareVie
   const anyFileHasDiffStatus = useRef<boolean>(false);
 
   const fileTree = useMemo(() => {
-    const allFiles = uniq(baseFileTree.concat(compareFileTree));
+    const allFiles = uniq<string>(baseFileTree.concat(compareFileTree));
     anyFileHasDiffStatus.current = false;
     // sort by diff status
     return !fileCompareDataByName
