+++ scopes/component/component/index.ts
@@ -1,7 +1,7 @@
 import { ComponentAspect } from './component.aspect';
 
 export type { GetComponentsOptions } from './get-component-opts';
-export type { UseComponentType } from './ui/use-component';
+export type { UseComponentType } from './ui/use-componentt';
 export { useComponentHost } from './host';
 export { Component, InvalidComponent } from './component';
 export { ComponentID } from '@teambit/component-id';
