+++ scopes/component/component/ui/index.ts
@@ -3,7 +3,7 @@ export { Component } from './component';
 export { ConsumeMethodSlot, ComponentMenu, VersionRelatedDropdowns } from './menu';
 export { ComponentModel, ComponentModelProps } from './component-model';
 export { ComponentContext, ComponentProvider } from './context';
-export { useComponent } from './use-component';
+export { useComponent } from './use-componentt';
 export { TopBarNav } from './top-bar-nav';
 export { componentIdFields, componentOverviewFields, componentFields } from './use-component-query';
 export { useIdFromLocation } from './use-component-from-location';
