+++ scopes/component/component/ui/use-component-from-location.tsx
