--- scopes/component/component/ui/use-component/index.ts
@@ -1,22 +1,0 @@
-export { useComponent } from './use-component';
-export {
-  componentIdFields,
-  componentOverviewFields,
-  componentFields,
-  componentFieldsWithLogs,
-  COMPONENT_QUERY_LOG_FIELDS,
-  GET_COMPONENT,
-  GET_COMPONENT_WITH_LOGS,
-} from './use-component.fragments';
-export { useIdFromLocation } from './use-component-from-location';
-export type {
-  LogFilter,
-  Filters,
-  UseComponentOptions,
-  ComponentQueryResult,
-  UseComponentType,
-  ComponentLogs,
-  ComponentLogsResult,
-} from './use-component.model';
-export { useComponentQuery } from './use-component-query';
-export { useComponentLogs } from './use-component-logs';
