--- scopes/component/component/ui/use-component/use-component-query.ts
@@ -1,169 +1,0 @@
-import { useMemo, useEffect, useRef } from 'react';
-import { useDataQuery } from '@teambit/ui-foundation.ui.hooks.use-data-query';
-import { ComponentID, ComponentIdObj } from '@teambit/component-id';
-import { ComponentDescriptor } from '@teambit/component-descriptor';
-import { ComponentModel } from '../component-model';
-import { ComponentQueryResult, Filters } from './use-component.model';
-import {
-  GET_COMPONENT,
-  SUB_COMPONENT_CHANGED,
-  SUB_COMPONENT_REMOVED,
-  SUB_SUBSCRIPTION_ADDED,
-} from './use-component.fragments';
-import { useComponentLogs } from './use-component-logs';
-import { ComponentError } from '../component-error';
-
-/** provides data to component ui page, making sure both variables and return value are safely typed and memoized */
-export function useComponentQuery(
-  componentId: string,
-  host: string,
-  filters?: Filters,
-  skip?: boolean
-): ComponentQueryResult {
-  const idRef = useRef(componentId);
-  idRef.current = componentId;
-  const variables = {
-    id: componentId,
-    extensionId: host,
-  };
-
-  const { data, error, loading, subscribeToMore } = useDataQuery(GET_COMPONENT, {
-    variables,
-    skip,
-    errorPolicy: 'all',
-  });
-
-  const { loading: loadingLogs, componentLogs: { logs } = {} } = useComponentLogs(componentId, host, filters, skip);
-
-  const rawComponent = data?.getHost?.get;
-
-  useEffect(() => {
-    // @TODO @Kutner fix subscription for scope
-    if (host !== 'teambit.workspace/workspace') {
-      return () => {};
-    }
-
-    const unsubAddition = subscribeToMore({
-      document: SUB_SUBSCRIPTION_ADDED,
-      updateQuery: (prev, { subscriptionData }) => {
-        const prevComponent = prev?.getHost?.get;
-        const addedComponent = subscriptionData?.data?.componentAdded?.component;
-
-        if (!addedComponent || prevComponent) return prev;
-
-        if (idRef.current === addedComponent.id.name) {
-          return {
-            ...prev,
-            getHost: {
-              ...prev.getHost,
-              get: addedComponent,
-            },
-          };
-        }
-
-        return prev;
-      },
-    });
-
-    const unsubChanges = subscribeToMore({
-      document: SUB_COMPONENT_CHANGED,
-      updateQuery: (prev, { subscriptionData }) => {
-        if (!subscriptionData.data) return prev;
-
-        const prevComponent = prev?.getHost?.get;
-        const updatedComponent = subscriptionData?.data?.componentChanged?.component;
-
-        const isUpdated = updatedComponent && ComponentID.isEqualObj(prevComponent?.id, updatedComponent?.id);
-
-        if (isUpdated) {
-          return {
-            ...prev,
-            getHost: {
-              ...prev.getHost,
-              get: updatedComponent,
-            },
-          };
-        }
-
-        return prev;
-      },
-    });
-
-    const unsubRemoval = subscribeToMore({
-      document: SUB_COMPONENT_REMOVED,
-      updateQuery: (prev, { subscriptionData }) => {
-        if (!subscriptionData.data) return prev;
-
-        const prevComponent = prev?.getHost?.get;
-        const removedIds: ComponentIdObj[] | undefined = subscriptionData?.data?.componentRemoved?.componentIds;
-        if (!prevComponent || !removedIds?.length) return prev;
-
-        const isRemoved = removedIds.some((removedId) => ComponentID.isEqualObj(removedId, prevComponent.id));
-
-        if (isRemoved) {
-          return {
-            ...prev,
-            getHost: {
-              ...prev.getHost,
-              get: null,
-            },
-          };
-        }
-
-        return prev;
-      },
-    });
-
-    return () => {
-      unsubChanges();
-      unsubAddition();
-      unsubRemoval();
-    };
-  }, []);
-
-  const idDepKey = rawComponent?.id
-    ? `${rawComponent?.id?.scope}/${rawComponent?.id?.name}@${rawComponent?.id?.version}}`
-    : undefined;
-
-  const id: ComponentID | undefined = useMemo(
-    () => (rawComponent ? ComponentID.fromObject(rawComponent.id) : undefined),
-    [idDepKey]
-  );
-
-  const componentError =
-    error && !data
-      ? new ComponentError(500, error.message)
-      : (!rawComponent && !loading && new ComponentError(404)) || undefined;
-
-  const component = useMemo(
-    () => (rawComponent ? ComponentModel.from({ ...rawComponent, host, logs }) : undefined),
-    [id?.toString(), logs]
-  );
-
-  const componentDescriptor = useMemo(() => {
-    const aspectList = {
-      entries: rawComponent?.aspects.map((aspectObject) => {
-        return {
-          ...aspectObject,
-          aspectId: aspectObject.id,
-          aspectData: aspectObject.data,
-        };
-      }),
-    };
-
-    return id ? ComponentDescriptor.fromObject({ id: id.toString(), aspectList }) : undefined;
-  }, [id?.toString()]);
-
-  return useMemo(() => {
-    return {
-      componentDescriptor,
-      component,
-      componentLogs: {
-        loading: loadingLogs,
-        logs,
-      },
-      error: componentError || undefined,
-      loading,
-    };
-  }, [host, component, componentDescriptor, componentError]);
-}
