--- scopes/component/component/ui/use-component/use-component.model.ts
@@ -1,45 +1,0 @@
-import { ComponentDescriptor } from '@teambit/component-descriptor';
-import { LegacyComponentLog } from '@teambit/legacy-component-log';
-import { ComponentError } from '../component-error';
-import { ComponentModel } from '../component-model';
-
-export type LogFilter = {
-  offset?: number;
-  limit?: number;
-  head?: string;
-  sort?: string;
-  takeHeadFromComponent?: boolean;
-};
-
-export type Filters = {
-  log?: LogFilter & { type?: string };
-  loading?: boolean;
-};
-
-export type UseComponentOptions = {
-  version?: string;
-  logFilters?: Filters;
-  customUseComponent?: UseComponentType;
-  skip?: boolean;
-};
-
-export type ComponentQueryResult = {
-  component?: ComponentModel;
-  componentDescriptor?: ComponentDescriptor;
-  componentLogs?: ComponentLogs;
-  loading?: boolean;
-  error?: ComponentError;
-};
-
-export type ComponentLogsResult = {
-  componentLogs?: ComponentLogs;
-  error?: ComponentError;
-  loading?: boolean;
-};
-
-export type ComponentLogs = {
-  logs?: LegacyComponentLog[];
-  loading?: boolean;
-};
-
-export type UseComponentType = (id: string, host: string, filters?: Filters, skip?: boolean) => ComponentQueryResult;
