--- scopes/component/component/ui/use-component/use-component.utils.ts
@@ -1,9 +1,0 @@
-export function getOffsetValue(offset, limit, backwards = false) {
-  if (offset !== undefined) {
-    return backwards ? -(offset + limit) : offset;
-  }
-  if (limit !== undefined) {
-    return 0;
-  }
-  return undefined;
-}
